- (snmp_impl.h): change pdu version from long to int.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@1429 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/snmp_impl.h
@@ -57,7 +57,7 @@ SOFTWARE.
 struct packet_info {
     u_char 	pdutype;
     snmp_ipaddr	source;
-    long	version;
+    int 	version;
     int		mp_model;
     int		sec_model;
     int		sec_level;
